+++ src/Model.h
@@ -12,11 +12,11 @@ class Model {
     static uint8_t maxWeights;
      
     uint8_t id;
-    float defaultW;
+    uint8_t defaultW;
     uint8_t emitGroups;
     HashMap<uint8_t, Weight*> *weights;
     
-    Model(uint8_t id, float defaultW, uint8_t emitGroups) : id(id), defaultW(defaultW), emitGroups(emitGroups) {
+    Model(uint8_t id, uint8_t defaultW, uint8_t emitGroups) : id(id), defaultW(defaultW), emitGroups(emitGroups) {
         weights = new HashMap<uint8_t, Weight*>(maxWeights);
         weights->setNullValue(NULL);
     }
@@ -25,25 +25,25 @@ class Model {
         delete weights;
     }
     
-    void put(Port *outgoing, Port *incoming, float weight) {
+    void put(Port *outgoing, Port *incoming, uint8_t weight) {
       _getOrCreate(outgoing, defaultW)->add(incoming, weight);
       _getOrCreate(incoming, defaultW)->add(outgoing, weight);
     }
     
-    void put(Port *outgoing, float w) {
+    void put(Port *outgoing, uint8_t w) {
       _getOrCreate(outgoing, w);
     }
     
-    void put(Connection *con, float w1, float w2) {
+    void put(Connection *con, uint8_t w1, uint8_t w2) {
       put(con->fromPort, w1);
       put(con->toPort, w2);
     }
     
-    void put(Connection *con, float w) {
+    void put(Connection *con, uint8_t w) {
       put(con, w, w);
     }
     
-    float get(Port *outgoing, Port *incoming) {
+    uint8_t get(Port *outgoing, Port *incoming) {
       if (outgoing == incoming) {
         return 0;
       }
@@ -54,7 +54,7 @@ class Model {
       return defaultW;
     }
     
-    Weight *_getOrCreate(Port *outgoing, float w) {
+    Weight *_getOrCreate(Port *outgoing, uint8_t w) {
         Weight *weight = weights->get(outgoing->id);
         if (weight == NULL) {
             weight = new Weight(w);
